Version 1.1.3-003 (tag: map_v1_1_3-003): Fixes for CUI-86 Map Component - Route Leg Step Maps.
MapDelegateImpl now implements:
- getInitialRouteLegStepMap
- manipulateRouteLegStepMap
- getInitialRouteMap (client side generated map support added).

Now depends on sdpCommon-manager 1.0-065.

Showcase updated to display basic route leg step links when a route summary
map is displayed. All route and leg step maps now work for both hi end
and classic maps. One known issue though: WhereisMobile contains
special code that tries to detect when a pan/zoom of a route summary
map returns to the original map. It is not perfect, though, and this
component does not (yet) cater to this.